Add solution in C for Problem 4
This is a brute-force approach and definitely needs optimisation
for larger numbers than in the problem specified, but for 3-digit
numbers it is fast enough.

I will hopefully come back to refactor this solution as soon as
I have found more mathematical approach to generate palindrome
products.

Addionally I changed the layout for the README and will use this
one from now on.