Add solution in C for Problem 13
Well, this is one clunky as solution for this problem, but I think
I did quite well in dealing with 50-digits numbers in C by using
arrays to represent and sum up all the numbers.

Nontheless, there is heaps of optimisation possible and hopefully
I will remember to come back later and refactor this a little.