Add solution for Problem 8 in C
Well, this was a nice exercise. No fancy math stuff going on here, just reading the
input file and sort of brute forcing my way to find the solution. ^^

I could probably make it a wee bit faster, but it's plenty 'fast' (however your own
definition is) to solve the problem. Could already check for zeros while reading the
numbers and then discard the current sequence if a 0 is found.