Add solution for Problem 9 in C
This was a fun problem and yet again made me realise how little I know
about math. ^^ Tried a couple different approaches and thanks to
http://www.maths.surrey.ac.uk/hosted-sites/R.Knott/Pythag/pythag.html
I was able to find a nice way to solve this.

I settled for the `m, n formula` to generate Pythagorean Triples, and
although it's sort of brute-force, because I'm looping through a sequence
of number, it does the trick.