Add solution in C for Problem 12
*phew* This was a tougher nut to crack, and I cheated a little. ^^

I set a limit to the number of triangle numbers being generated and it
takes roughly 15 seconds to generate and parse the Eratosthenes Sieve
with the upper bound of the largest triangular number being used.

Afterwards I start at the end and prime factorise those triangular numbers
and check the number of divisors. It's a little cheating, because I hoped
that I would encounter the first number with the number of devisors over
500. ^^ Payed off and I solved the problem. :P

Was a nice problem to solve. :)